requirements.txt
alabaster==0.7.12
appdirs==1.4.4
argh==0.26.2
asgiref==3.2.7
astroid==2.4.1
Babel==2.8.0
beautifulsoup4==4.9.1
Bootstrap-Flask==1.2.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
distlib==0.3.0
Django==3.0.6
django-appconf==1.0.4
django-google-maps==0.12.1
django-phonenumber-field==4.0.0
Django-Select2==5.8.10
docutils==0.16
dominate==2.5.1
dwebsocket==0.5.12
Faker==4.0.2
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==0.9.0
Flask-MySQL==1.5.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
google==2.0.3
google-api-core==1.19.1
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
livereload==2.6.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
packaging==20.4
pathtools==0.1.2
phonenumbers==8.12.4
Pillow==7.1.2
pipenv==2020.6.2
port-for==0.3.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
restructuredtext-lint==1.3.0
rsa==4.0
rstcheck==3.3.1
six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.0.3
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.15
sqlparse==0.3.1
suit==2.0.2
text-unidecode==1.3
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
visitor==0.1.3
watchdog==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.2.1
zipp==3.1.0
